requirements.txt
torch==2.9.1
torchvision==0.24.1
transformers==4.57.3
accelerate==1.12.0
datasets==4.4.1
diffusers==0.36.0
scikit-learn==1.7.2
numpy==2.2.6
pandas==2.3.3
matplotlib==3.10.8
seaborn==0.13.2
scipy==1.15.3
Pillow==12.0.0
tqdm==4.67.1
